Move RegistrationForm to model package
